[Align] TrackPar and TrackParCov as template
Converts TrackPar and TrackParCov classes into template classes to allow
for a single and double precision floating point version. No changes in
program semantics.
